pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "ey-data-platform"
version = "0.1.0"
description = "Production data platform for EY technical test"
requires-python = ">=3.9"
dependencies = [
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "requests>=2.25.0",
    "google-cloud-storage>=2.0.0",
    "google-cloud-bigquery>=3.0.0",
    "google-cloud-pubsub>=2.0.0",
    "schedule>=1.2.0",  # For batch job scheduling
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "black>=23.0.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]